Subclass reorganiztion, callable and ordered matrices
